+++ packages/bpk-component-drawer/src/BpkDrawerContent-test.js
@@ -36,7 +36,6 @@ describe('BpkDrawerContent', () => {
           onCloseAnimationComplete={jest.fn()}
           closeLabel="Close"
           dialogRef={jest.fn()}
-          isIphone={false}
         >
           Drawer content
         </BpkDrawerContent>,
@@ -56,7 +55,6 @@ describe('BpkDrawerContent', () => {
           onCloseAnimationComplete={jest.fn()}
           closeLabel="Close"
           dialogRef={jest.fn()}
-          isIphone={false}
         >
           Drawer content
         </BpkDrawerContent>,
@@ -76,7 +74,6 @@ describe('BpkDrawerContent', () => {
           onCloseAnimationComplete={jest.fn()}
           closeLabel="Close"
           dialogRef={jest.fn()}
-          isIphone={false}
         >
           Drawer content
         </BpkDrawerContent>,
@@ -96,6 +93,25 @@ describe('BpkDrawerContent', () => {
           closeLabel="Close"
           dialogRef={jest.fn()}
           hideTitle
+        >
+          Drawer content
+        </BpkDrawerContent>,
+      )
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it('should render correctly with arbitrary attributes', () => {
+    const tree = renderer
+      .create(
+        <BpkDrawerContent
+          id="my-drawer"
+          title="Drawer title"
+          onClose={jest.fn()}
+          onCloseAnimationComplete={jest.fn()}
+          closeLabel="Close"
+          dialogRef={jest.fn()}
+          data-arbitrary="prop"
         >
           Drawer content
         </BpkDrawerContent>,